saving of images outside of the training and now added checkpoint loading to apply validation on its own

+++ lesseg_unet/training.py
@@ -9,41 +9,13 @@ import numpy as np
 import monai
 import torch
 from monai.metrics import DiceMetric
-from monai.data import Dataset
 from monai.transforms import (
     Activations,
     AsDiscrete,
     Compose,
 )
-from monai.visualize import plot_2d_or_3d_image
 from torch.utils.tensorboard import SummaryWriter
-from lesseg_unet import data_loading, net, utils, transformations
-
-
-def init_training_data(img_path_list: Sequence,
-                       seg_path_list: Sequence,
-                       img_pref: str = None,
-                       transform_dict=None,
-                       train_val_percentage: float = 75) -> Tuple[monai.data.Dataset, monai.data.Dataset]:
-    logging.info('Listing input files to be loaded')
-    train_files, val_files = data_loading.create_file_dict_lists(img_path_list, seg_path_list, img_pref,
-                                                                 train_val_percentage)
-    logging.info('Create transformations')
-    train_img_transforms = transformations.segmentation_train_transformd(transform_dict)
-    val_img_transforms = transformations.segmentation_val_transformd(transform_dict)
-    # define dataset, data loader
-    logging.info('Create training monai datasets')
-    train_ds = Dataset(train_files, transform=train_img_transforms)
-
-    # define dataset, data loader
-    logging.info('Create validation actual monai datasets')
-    val_ds = Dataset(val_files, transform=val_img_transforms)
-    # We check if both the training and validation dataloaders can be created and used without immediate errors
-    logging.info('Checking data loading')
-    data_loading.data_loader_checker_first(train_ds, 'training')
-    data_loading.data_loader_checker_first(val_ds, 'validation')
-    logging.info('Init training done.')
-    return train_ds, val_ds
+from lesseg_unet import net, utils, data_loading
 
 
 def training_loop(img_path_list: Sequence,
@@ -61,7 +33,8 @@ def training_loop(img_path_list: Sequence,
     else:
         device = torch.device(device)
     val_output_affine = utils.nifti_affine_from_dataset(img_path_list[0])
-    train_ds, val_ds = init_training_data(img_path_list, seg_path_list, img_pref, transform_dict=transform_dict)
+    train_ds, val_ds = data_loading.init_training_data(img_path_list, seg_path_list, img_pref,
+                                                       transform_dict=transform_dict, train_val_percentage=75)
     train_loader = data_loading.create_training_data_loader(train_ds, batch_size, dataloader_workers)
     val_loader = data_loading.create_validation_data_loader(val_ds, dataloader_workers=dataloader_workers)
     model = net.create_unet_model(device, net.default_unet_hyper_params)
@@ -114,6 +87,7 @@ def training_loop(img_path_list: Sequence,
     best_metric_epoch = -1
     epoch_loss_values = list()
     metric_values = list()
+    val_save_thr = 1
     """
     Measure tracking init
     """
@@ -198,23 +172,23 @@ def training_loop(img_path_list: Sequence,
                     val_score_list.append(value.item())
                     metric_count += len(value)
                     metric_sum += value.item() * len(value)
-                    if best_metric > 0.7:
-                        if value.item() * len(value) > 0.7:
-                            print('Saving good image #{}'.format(img_count))
-                            img_count += 1
-                            if img_count < num_nifti_save:
-                                utils.save_img_lbl_seg_to_png(
-                                    inputs, val_images_dir, 'validation_img_{}'.format(img_count), labels, outputs)
-                                utils.save_img_lbl_seg_to_nifti(
-                                    inputs, labels, outputs, val_images_dir, val_output_affine, img_count)
-                        if value.item() * len(value) < 0.1:
-                            trash_count += 1
-                            print('Saving trash image #{}'.format(trash_count))
-                            if trash_count < num_nifti_save:
-                                utils.save_img_lbl_seg_to_png(
-                                    inputs, trash_val_images_dir, 'trash_img_{}'.format(trash_count), labels, outputs)
-                                utils.save_img_lbl_seg_to_nifti(
-                                    inputs, labels, outputs, trash_val_images_dir, val_output_affine, trash_count)
+                    # if best_metric > val_save_thr:
+                    #     if value.item() * len(value) > val_save_thr:
+                    #         img_count += 1
+                    #         if img_count < num_nifti_save:
+                    #             print('Saving good image #{}'.format(img_count))
+                    #             utils.save_img_lbl_seg_to_png(
+                    #                 inputs, val_images_dir, 'validation_img_{}'.format(img_count), labels, outputs)
+                    #             utils.save_img_lbl_seg_to_nifti(
+                    #                 inputs, labels, outputs, val_images_dir, val_output_affine, img_count)
+                    #     if value.item() * len(value) < 0.1:
+                    #         trash_count += 1
+                    #         if trash_count < num_nifti_save:
+                    #             print('Saving trash image #{}'.format(trash_count))
+                    #             utils.save_img_lbl_seg_to_png(
+                    #                 inputs, trash_val_images_dir, 'trash_img_{}'.format(trash_count), labels, outputs)
+                    #             utils.save_img_lbl_seg_to_nifti(
+                    #                 inputs, labels, outputs, trash_val_images_dir, val_output_affine, trash_count)
                 metric = metric_sum / metric_count
                 metric_values.append(metric)
                 median = np.median(np.array(val_score_list))
@@ -240,21 +214,24 @@ def training_loop(img_path_list: Sequence,
                 if metric > best_metric:
                     best_metric = metric
                     best_metric_epoch = epoch + 1
-                    torch.save(model.state_dict(),
-                               Path(output_dir,
-                                    'best_metric_model_segmentation3d_array_epo_{}.pth'.format(best_metric_epoch)))
+                    # torch.save(model.state_dict(),
+                    #            Path(output_dir,
+                    #                 'best_metric_model_segmentation3d_array_epo_{}.pth'.format(best_metric_epoch)))
+                    utils.save_checkpoint(model, epoch + 1, optimizer, output_dir,
+                                          'best_metric_model_segmentation3d_array_epo.pth')
                     print('saved new best metric model')
                     writer.add_scalar('val_best_mean_dice', metric, epoch + 1)
                     df.at[epoch + 1, 'val_best_mean_dice'] = metric
-                    print('Replacing best images')
-                    for f in best_val_images_dir.iterdir():
-                        os.remove(f)
-                    for f in best_trash_images_dir.iterdir():
-                        os.remove(f)
-                    for f in val_images_dir.iterdir():
-                        shutil.copy(f, best_val_images_dir)
-                    for f in trash_val_images_dir.iterdir():
-                        shutil.copy(f, best_trash_images_dir)
+                    if best_metric > val_save_thr:
+                        print('Replacing best images')
+                        for f in best_val_images_dir.iterdir():
+                            os.remove(f)
+                        for f in best_trash_images_dir.iterdir():
+                            os.remove(f)
+                        for f in val_images_dir.iterdir():
+                            shutil.copy(f, best_val_images_dir)
+                        for f in trash_val_images_dir.iterdir():
+                            shutil.copy(f, best_trash_images_dir)
                     # plot the last model output as GIF image in TensorBoard with the corresponding image and label
                     # plot_2d_or_3d_image(val_images, epoch + 1, writer, index=0, tag="image")
                     # plot_2d_or_3d_image(val_labels, epoch + 1, writer, index=0, tag="label")
@@ -269,7 +246,7 @@ def training_loop(img_path_list: Sequence,
                     )
                 )
 
-                utils.save_checkpoint(model, epoch + 1, optimizer, output_dir)
+                # utils.save_checkpoint(model, epoch + 1, optimizer, output_dir)
     df.to_csv(Path(output_dir, 'perf_measures.csv'), columns=perf_measure_names)
     print(f'train completed, best_metric: {best_metric:.4f} at epoch: {best_metric_epoch}')
     writer.close()
